Replace summary in README.md
Replaces the summary content in the README.md file and updates the README.md file so that the summary section is managed manually. The existing summary originates from the OpenAPI spec, which is not tailored to address the SDK use case. The updated summary also includes the boilerplate language that is approved for preview features.

The links to the Cribl as Code docs are speculative until two PRs in the cribl-doc repo are merged:

- https://bitbucket.org/cribl/cribl-doc/pull-requests/5759
- https://bitbucket.org/cribl/cribl-doc/pull-requests/5696